+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Entries without a reference link were developed exclusively within the [batumibi
 - [#14820](https://github.com/opencart/opencart/pull/14820) - Event list: always allow for viewing event details.
 - [#14891](https://github.com/opencart/opencart/pull/14891) - To collect cache files, do it only if the random has fallen to the cleanup.
 - [#14915](https://github.com/opencart/opencart/pull/14915) - Add `OPTIMIZE TABLE` to `db.php`.
+- [#15326](https://github.com/opencart/opencart/pull/15326) - Use the composer-generated autoload for the 3rd-party packages
 - [0a8b25a](https://github.com/opencart/opencart/commit/0a8b25aaedae97e1b21784263527941448724693) - Added some additional form text.
 - [8996eb3](https://github.com/opencart/opencart/commit/8996eb371620c3c7a453872d39a3a815562690f3) - `partially used` - Improved `admin/controller/marketplace/installer.php`.
 - [062ba0e](https://github.com/opencart/opencart/commit/062ba0ee243cc707b51b95e56b69b7f07cda5f7f), [85e8ddb](https://github.com/opencart/opencart/commit/85e8ddb82916cc48a4dcb30fcfcd4a1a9ea9add0) - `partially used` - Updated language methods.